+++ test/manual/test_minuit_fitters.cpp
@@ -27,10 +27,10 @@ void compareRootMinuit(void)
     double maxTime = 0.002;
 
     size_t numCfEvents  = 1000000;
-    size_t numDcsEvents = PullStudyHelpers::numDCSDecays(numCfEvents, phaseSpaceParams, maxTime);
+    double numDcsEvents = PullStudyHelpers::numDCSDecays(numCfEvents, phaseSpaceParams, maxTime);
 
     SimulatedDecays MyDecays = SimulatedDecays(maxTime, phaseSpaceParams);
-    MyDecays.findDcsDecayTimes(numDcsEvents);
+    MyDecays.findDcsDecayTimes((size_t)numDcsEvents);
     MyDecays.findCfDecayTimes(numCfEvents);
 
     // Define some time bins
